MAINTENANCE: some cleanup and refactor of Resource 'module'
I originally thought Project would be enough to refactor, but I was wrong.

Note: you'll notice the reorganization of the Events. This is because the latest version of Backbone.Radio has gotten rid of 'Event' and 'Command'. So, I thought I'd reorganize this file for when we move to the latest version of Radio.