FEATURE: server wait timeout
if x milliseconds go by and we're still waiting for responses from the server, we fire an event in the client. This is useful when we're asking a lot of the server. For example, importing Workflows could take over a second.
In that case, the WorkflowBuilder can temporarily disable mouse events on the GUI and control, just in case the user gets antsy.