FIX: #57
Aside from documenting all Events, the major improvement was decoupling the ControllerWorkflowBuilder from the GUI. Rather, the ControllerWorkflowBuilder is now essentially a collection of functions that, when given models it manipulates them without having to remember their state.
It's an event proxy that acts as a bridge between the GUI and the actual model controllers. The end result is:

- reduced code
- well-defined events
- decoupling of GUI from model controllers