FEATURE: Radio response events for Project requests
On success or failure of a Project save/creation/deletion the Project controller sends out a Radio event. These new events are a kind of "model" that will be employed more in the client. This way I can spawn a "Please wait...doing something..." message when the client is doing some heavy lifting, so the user doesn't start to worry that something has gone wrong.