FIX: #1 FIX: #2
Configuration now held in configuration.json. This file is requested by the client before any other initialization. This means that we can change the configuration "on the fly" without having to redist in production. (Of course, the webpage has to be reloaded, but that's expected).

Also, there are defaults for all config settings EXCEPT server info and amdin name/email. See configuration.json.example